Basic save save and checkpoint done
I spent to long trying to find some super difficult error that actually wasn't difficult at all. I didn't specify the name of a save file on the main menu scene. This was breaking the build which started on the main menu and not on level 1:() The code needs some clean up though.